Adding JUnit tests for Engine and introducing CLOCK_MAX_TIME

+++ app/src/main/java/daw/engine/RPEngine.java
@@ -8,9 +8,9 @@ public interface RPEngine {
 	
 	void stop();
 	
-	void setPlaybackTime(long time);
+	void setPlaybackTime(Double time);
 	
-	long getPlaybackTime();
+	Double getPlaybackTime();
 	
 	boolean isPaused();
